Revert "SEACAS: Sync with seacas; nemesis large model changes"
This reverts commit 3a547772488f20a0a909eb0431daf6e57ab3a2b3.

+++ packages/seacas/libraries/nemesis/ne_gcmp.c
@@ -81,7 +81,7 @@ int ne_get_cmap_params(int  neid,
   char   *func_name="ne_get_cmap_params";
 
   size_t  cnt, num_n_comm_maps, num_e_comm_maps, start[1], count[1];
-  int64_t cmap_info_idx[2], cmap_data_idx[2];
+  size_t  cmap_info_idx[2], cmap_data_idx[2];
   int     nmstat;
   int     status, map_idx, varid, dimid;
